✅ ボタンコンポーネントのテストをApp.test.jsxから移動
